chore(sdk): azure refactor resources to app level
